split into async and blocking features (#36)
* split into async and blocking features

* update Cargo.toml file

* fix example
