+++ fang_examples/simple_worker/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2018"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-fang = { path = "../../" }
+fang = { path = "../../" , features = ["blocking"]}
 serde = { version = "1.0", features = ["derive"] }
 signal-hook = "0.3.10"
 dotenv = "0.15.0"
