+++ src/lib.rs
@@ -1,29 +1,20 @@
-// #![allow(clippy::nonstandard_macro_braces)]
 #![allow(clippy::extra_unused_lifetimes)]
 
-// pub mod error;
-// pub mod executor;
-// pub mod queue;
-// pub mod scheduler;
-// pub mod schema;
-// pub mod worker_pool;
-
-// pub use error::FangError;
-// pub use executor::*;
-// pub use queue::*;
-// pub use scheduler::*;
-// pub use schema::*;
-// pub use worker_pool::*;
-
 #[macro_use]
+#[cfg(feature = "blocking")]
 extern crate diesel;
 
 #[doc(hidden)]
+#[cfg(feature = "blocking")]
 pub use diesel::pg::PgConnection;
+
 #[doc(hidden)]
 pub use typetag;
 
-pub mod sync;
-pub use sync::*;
+#[cfg(feature = "blocking")]
+pub mod blocking;
+#[cfg(feature = "blocking")]
+pub use blocking::*;
 
+#[cfg(feature = "asynk")]
 pub mod asynk;
